requirements.txt
blinker==1.7.0
Bottleneck @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/bottleneck_1699237772214/work
click==8.1.7
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.0
fsspec==2024.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numexpr @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/numexpr_1699237935415/work
numpy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_cfbbnu3tzv/croot/numpy_and_numpy_base_1704311724719/work/dist/numpy-1.26.3-cp312-cp312-macosx_11_0_arm64.whl#sha256=e066d59974be230052477e95f2633cdef4c4987c2f32c15291d3f4b184022bdc
opencv-python==4.9.0.80
pandas @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_82r7p4atwy/croot/pandas_1702318000240/work/dist/pandas-2.1.4-cp312-cp312-macosx_11_0_arm64.whl#sha256=142530670b1e4125a2466ecf7521689e9e72ac90d1e53a96b5ce39180a797870
pillow==10.2.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pytz_1699238107851/work
setuptools==68.2.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy==1.12
torch==2.2.0
typing_extensions==4.9.0
tzdata @ file:///croot/python-tzdata_1690578112552/work
waitress==3.0.0
Werkzeug==3.0.2
wheel==0.41.2
